REFACTOR: Rename contracts to types
