+++ src/conversion/index.ts
@@ -1,2 +1,2 @@
-export * from './contracts';
+export * from './types';
 export * from './azure';
